Change dts interface, adjust and refactor voltage divider
All changes in the files started with changes in
dts/bindings/irnas,voltage-divider.yaml.
The way how resistor values are specified changed, additional property
for measuring internal vdd was added.
Other changes:
- In drivers source files the devicetree macros had to be adjusted due to
that (they had to change anyway as the used label properties, which were
removed in the ncs 2.2.0).
- The rest of the changes in the driver source files are related
to validation of device tree settings and general cleanup.
- Header includes in all files were updated to include zephyr path.
- Driver is now automatically compiled with the application, if a node
  with 'compatible' property is set to "irnas,voltage-divider".